Fixed sticky notes sometimes flicker.
This was caused by tips not having their source as an ancestor
in storage configuration.
To address this, mouseout in Source now checks the data-rit-id of the
relatedTarget to make sure one is indeed leaving the source/tip.
To make this reliable data-rit-id was moved from tip to Location,
which also simplifies the code overall because individual tips
(Ballon, Cloud, ...) do not need to care about data-rit-id anymore